src/main/java/com/galvanize/useraccounts/request/UpdateUserRequest.java
package com.galvanize.useraccounts.request;

public class UpdateUserRequest {

    private String firstName;
    private String lastName;
    private String password;
    private String email;
    private String address;
    private String creditCard;
    private boolean verify;

    public UpdateUserRequest() {}

    public UpdateUserRequest(String firstName, String lastName, String password, String email, String address, String creditCard, boolean verify) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.password = password;
        this.email = email;
        this.address = address;
        this.creditCard = creditCard;
        this.verify = verify;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getCreditCard() {
        return creditCard;
    }

    public void setCreditCard(String creditCard) {
        this.creditCard = creditCard;
    }

    public boolean isVerify() {
        return verify;
    }

    public void setVerify(boolean verify) {
        this.verify = verify;
    }
}
